move upload route to routes
