+++ .prettierrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+  bracketSpacing: true,
+  tabWidth: 2,
+  singleQuote: true,
+  trailingComma: 'es5',
+  importOrderSeparation: true,
+  importOrderSortSpecifiers: true,
+};
